git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/verify-keep-going@1465596 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svn/file-merge.c
@@ -367,7 +367,7 @@ prepare_line_for_display(const char *line, apr_pool_t *pool)
 {
   svn_stringbuf_t *buf = svn_stringbuf_create(line, pool);
   size_t width;
-  int line_width = LINE_DISPLAY_WIDTH;
+  size_t line_width = LINE_DISPLAY_WIDTH;
   apr_pool_t *iterpool;
 
   /* Trim EOL. */
